modified:   ESXi-Scripts/ps-env-setup.ps1 	modified:   dot_files/env_setup.sh

+++ dot_files/env_setup.sh
@@ -1,15 +1,40 @@
 #!/bin/bash
 
+#
+#   Setting up bashrc
+#
 echo "Backing up old .bashrc and linking new"
-mv ~/.bashrc ~/.bashrc.old
-ln -s ~/Home-Scripts/dot_files/bashrc ~/.bashrc
-
+if [ ! -L ~/.bashrc ]
+then
+    mv ~/.bashrc ~/.bashrc.old
+    ln -s ~/Home-Scripts/dot_files/bashrc ~/.bashrc
+fi
+#
+#   Setting up tmux env
+#
 echo "Setting up tmux env"
-ln -s ~/Home-Scripts/dot_files/tmux.conf ~/.tmux.conf
+if [ ! -L ~/.tmux.conf ]
+then
+    mv ~/.tmux.conf ~/.tmux.conf.old
+    ln -s ~/Home-Scripts/dot_files/tmux.conf ~/.tmux.conf
+    if [ -f ~/.tmux ]
+    then
+        mv ~/.tmux ~/.tmux.old
+    fi
+fi
 git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 echo "Run 'Ctrl-B Shft-I' from withing tmux to enable plugins"
-
+#
+#   Setting up vimrc
+#
 echo "Setting up awesome vimrc"
+if [ -f ~/.vimrc ]
+then
+    mv ~/.vimrc ~/.vimrc.old
+fi
 git clone --depth=1 https://github.com/amix/vimrc.git ~/.vim_runtime
 sh ~/.vim_runtime/install_awesome_vimrc.sh
-. ~/.bashrc
+#
+#   Refresh Env (might require a logout)
+#
+source ~/.bashrc
